PythonFiles/ML code/scripts/requirements_test.txt
absl-py==0.7.0
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
boto==2.49.0
boto3==1.9.98
botocore==1.12.98
bz2file==0.98
chardet==3.0.4
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
gast==0.2.2
gensim==3.7.1
grpcio==1.18.0
h5py==2.9.0
html5lib==0.9999999
idna==2.8
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.4
mock==2.0.0
networkx==2.2
nibabel==2.3.3
nltk==3.4
numpy==1.16.1
opencv-python==4.0.0.21
pandas==0.24.1
pbr==5.1.3
Pillow==5.4.1
protobuf==3.6.1
pydicom==1.2.2
pyparsing==2.3.1
Pyphen==0.9.5
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
radiomics==0.1
repoze.lru==0.7
requests==2.21.0
s3transfer==0.2.0
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
SimpleITK==1.2.0
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.0
stop-words==2018.7.23
subprocess32==3.5.3
tensorboard==1.8.0
tensorflow==1.7.0
termcolor==1.1.0
textstat==0.5.4
Theano==1.0.4
toolz==0.9.0
tqdm==4.31.1
urllib3==1.24.1
Werkzeug==0.14.1
